app/requirements.txt
anyio==3.6.2
azure-cognitiveservices-vision-contentmoderator==1.0.0
azure-common==1.1.28
azure-core==1.26.3
azure-cosmos==4.3.0
azure-identity==1.12.0
black==23.1.0
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.0.1
click==8.1.3
cloudinary==1.31.0
cryptography==39.0.1
dnspython==2.3.0
EditorConfig==0.12.3
email-validator==1.3.1
Faker==16.6.1
fast-query-parsers==0.3.0
filetype==1.2.0
flake8==6.0.0
geojson==2.5.0
h11==0.14.0
httpcore==0.16.3
httpx==0.23.3
idna==3.4
isodate==0.6.1
jsbeautifier==1.14.7
Mako==1.2.4
markdown-it-py==2.1.0
MarkupSafe==2.1.2
mccabe==0.7.0
mdurl==0.1.2
msal==1.21.0
msal-extensions==1.0.0
msgspec==0.12.0
msrest==0.7.1
multidict==6.0.4
mypy-extensions==1.0.0
oauthlib==3.2.2
packaging==23.0
pathspec==0.11.0
picologging==0.9.1
Pillow==9.4.0
platformdirs==2.6.2
portalocker==2.7.0
pycodestyle==2.10.0
pycparser==2.21
pydantic==1.10.4
pydantic-factories==1.17.1
pydantic-openapi-schema==1.5.1
pyflakes==3.0.1
Pygments==2.14.0
PyJWT==2.6.0
python-dateutil==2.8.2
python-dotenv==0.21.1
PyYAML==6.0
requests==2.28.2
requests-oauthlib==1.3.1
rfc3986==1.5.0
rich==13.3.1
six==1.16.0
sniffio==1.3.0
starlite==1.51.0
typing_extensions==4.4.0
urllib3==1.26.14
uvicorn==0.20.0
